+++ src/test/java/org/folio/sidecar/service/ErrorHandlerTest.java
@@ -28,6 +28,7 @@ import org.apache.http.ParseException;
 import org.folio.sidecar.exception.EgressUnauthorizedException;
 import org.folio.sidecar.exception.EntitlementsNotLoadedException;
 import org.folio.sidecar.exception.KeycloakUnhandledAuthorizationException;
+import org.folio.sidecar.exception.ModUsersKeycloakTargetNotResolvedException;
 import org.folio.sidecar.exception.TenantNotEnabledException;
 import org.folio.sidecar.model.error.ErrorResponse;
 import org.folio.sidecar.support.TestUtils;
@@ -242,6 +243,20 @@ class ErrorHandlerTest {
 
     assertThat(responseCaptor.getValue())
       .isEqualTo(TestUtils.minify(TestUtils.readString("json/entitlements-not-loaded-error-timeout-cause.json")));
+    assertThat(responseStatusCaptor.getValue()).isEqualTo(SC_SERVICE_UNAVAILABLE);
+    verify(routingContext.response()).putHeader(RETRY_AFTER, "5");
+    verify(jsonConverter).toJson(any(ErrorResponse.class));
+    verify(sidecarSignatureService).removeSignature(routingContext);
+  }
+
+  @Test
+  void sendErrorResponse_positive_modUsersKeycloakTargetNotResolvedError() {
+    var routingContext = routingContext();
+
+    errorHandler.sendErrorResponse(routingContext, new ModUsersKeycloakTargetNotResolvedException("test-tenant"));
+
+    assertThat(responseCaptor.getValue())
+      .isEqualTo(TestUtils.minify(TestUtils.readString("json/mod-users-keycloak-target-not-resolved-error.json")));
     assertThat(responseStatusCaptor.getValue()).isEqualTo(SC_SERVICE_UNAVAILABLE);
     verify(routingContext.response()).putHeader(RETRY_AFTER, "5");
     verify(jsonConverter).toJson(any(ErrorResponse.class));